32/64 bit problem fixed
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1187 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/ui/QsiBox.cpp
@@ -855,7 +855,7 @@ QsiBlock* QsiBox::lookup(QGraphicsItem* gItemP)
 			}
 
 			LM_T(LmtBlockLookup, ("Comparing pressed gItem '%p' to %s '%s' proxy '%p'", gItemP, qbP->typeName(), qbP->name, qbP->proxy));
-			if (((int) gItemP) == (((int) qbP->proxy) + 8))
+			if (((long) gItemP) == (((long) qbP->proxy) + 8))
 			{
 				LM_T(LmtBlockLookup, ("Found proxy %s '%s' in box '%s'", qbP->typeName(), qbP->name, name));
 				return qbP;
